move os video funtions to new file, removed postmessage to video
